eth2/python/requirements.txt
# pip freeze export
web3==5.11.0
python-dotenv